chore(create-append-anything): move rules to dedicated module
