+++ lib/features/create-append-anything/AppendRules.js
@@ -1,0 +1,71 @@
+import {
+  find,
+} from 'min-dash';
+
+import inherits from 'inherits-browser';
+
+import {
+  is,
+  getBusinessObject
+} from '../../util/ModelUtil';
+
+import {
+  isAny
+} from '../modeling/util/ModelingUtil';
+
+import RuleProvider from 'diagram-js/lib/features/rules/RuleProvider';
+
+
+/**
+ * Append anything modeling rules
+ */
+export default function AppendRules(eventBus) {
+  RuleProvider.call(this, eventBus);
+}
+
+inherits(AppendRules, RuleProvider);
+
+AppendRules.$inject = [
+  'eventBus'
+];
+
+AppendRules.prototype.init = function() {
+  this.addRule('shape.append', function(context) {
+
+    var source = context.element;
+
+    const businessObject = getBusinessObject(source);
+
+    if (isAny(source, [
+      'bpmn:EndEvent',
+      'bpmn:Group',
+      'bpmn:TextAnnotation',
+      'bpmn:SequenceFlow',
+      'bpmn:Lane',
+      'bpmn:Participant',
+      'bpmn:DataStoreReference',
+      'bpmn:DataObjectReference'
+    ])) {
+      return false;
+    }
+
+    if (is(source, 'bpmn:IntermediateThrowEvent') && hasEventDefinition(source, 'bpmn:LinkEventDefinition')) {
+      return false;
+    }
+
+    if (is(source, 'bpmn:SubProcess') && businessObject.triggeredByEvent) {
+      return false;
+    }
+  });
+
+};
+
+
+// helpers //////////////
+function hasEventDefinition(element, eventDefinition) {
+  var bo = getBusinessObject(element);
+
+  return !!find(bo.eventDefinitions || [], function(definition) {
+    return is(definition, eventDefinition);
+  });
+}
