+++ lib/features/create-append-anything/index.js
@@ -9,6 +9,7 @@ import CreateAppendEditorActions from './CreateAppendEditorActions';
 import CreateAppendKeyboardBindings from './CreateAppendKeyboardBindings';
 import AppendMenuProvider from './AppendMenuProvider';
 import AppendContextPadProvider from './AppendContextPadProvider';
+import AppendRules from './AppendRules';
 
 export default {
   __depends__: [
@@ -23,12 +24,14 @@ export default {
     'createAppendEditorActions',
     'createAppendKeyboardBindings',
     'appendMenuProvider',
-    'appendContextPadProvider'
+    'appendContextPadProvider',
+    'appendRules'
   ],
   createMenuProvider: [ 'type', CreateMenuProvider ],
   createPaletteProvider: [ 'type', CreatePaletteProvider ],
   createAppendEditorActions: [ 'type', CreateAppendEditorActions ],
   createAppendKeyboardBindings: [ 'type', CreateAppendKeyboardBindings ],
   appendMenuProvider: [ 'type', AppendMenuProvider ],
-  appendContextPadProvider: [ 'type', AppendContextPadProvider ]
+  appendContextPadProvider: [ 'type', AppendContextPadProvider ],
+  appendRules: [ 'type', AppendRules ]
 };
