+++ lib/features/rules/BpmnRules.js
@@ -169,11 +169,6 @@ BpmnRules.prototype.init = function() {
 
     return canCopy(elements, element);
   });
-
-  this.addRule('shape.append', function(context) {
-    var element = context.element;
-    return canAppend(element);
-  });
 };
 
 BpmnRules.prototype.canConnectMessageFlow = canConnectMessageFlow;
@@ -804,32 +799,6 @@ function canResize(shape, newBounds) {
   }
 
   return false;
-}
-
-function canAppend(source) {
-
-  const businessObject = getBusinessObject(source);
-
-  if (isAny(source, [
-    'bpmn:EndEvent',
-    'bpmn:Group',
-    'bpmn:TextAnnotation',
-    'bpmn:SequenceFlow',
-    'bpmn:Lane',
-    'bpmn:Participant',
-    'bpmn:DataStoreReference',
-    'bpmn:DataObjectReference'
-  ])) {
-    return false;
-  }
-
-  if (is(source, 'bpmn:IntermediateThrowEvent') && hasEventDefinition(source, 'bpmn:LinkEventDefinition')) {
-    return false;
-  }
-
-  if (is(source, 'bpmn:SubProcess') && businessObject.triggeredByEvent) {
-    return false;
-  }
 }
 
 /**
