+++ test/spec/features/rules/BpmnRulesSpec.js
@@ -2181,73 +2181,4 @@ describe('features/modeling/rules - BpmnRules', function() {
 
   });
 
-
-  describe('element append', function() {
-
-    var testXML = require('../../../fixtures/bpmn/simple.bpmn');
-
-    beforeEach(bootstrapModeler(testXML, { modules: testModules }));
-
-
-    it('should not allow for given element types', inject(function(elementFactory, rules) {
-
-      // given
-      var types = [
-        'bpmn:EndEvent',
-        'bpmn:Group',
-        'bpmn:TextAnnotation',
-        'bpmn:SequenceFlow',
-        'bpmn:Lane',
-        'bpmn:Participant',
-        'bpmn:DataStoreReference',
-        'bpmn:DataObjectReference'
-      ];
-
-      // when
-
-      var results = types.map(function(type) {
-        var element = elementFactory.createShape({ type: type });
-        return rules.allowed('shape.append', { element });
-      });
-
-      // then
-      results.forEach(function(result) {
-        expect(result).to.be.false;
-      });
-    }));
-
-
-    it('should not allow for event subprocess', inject(function(elementFactory, rules) {
-
-      // given
-      var element = elementFactory.createShape({ type: 'bpmn:SubProcess', triggeredByEvent: true });
-
-      // when
-
-      var result = rules.allowed('shape.append', { element });
-
-      // then
-      expect(result).to.be.false;
-    }));
-
-
-    it('should not allow for link intermediate throw event', inject(function(elementFactory, rules) {
-
-      // given
-      var element = elementFactory.createShape({
-        type: 'bpmn:IntermediateThrowEvent',
-        cancelActivity: false,
-        eventDefinitionType: 'bpmn:LinkEventDefinition'
-      });
-
-      // when
-
-      var result = rules.allowed('shape.append', { element });
-
-      // then
-      expect(result).to.be.false;
-    }));
-
-  });
-
 });
